Cleanup Overtype overrides
