+++ package.json
@@ -79,7 +79,7 @@
     "axios": "^1.17.0",
     "laravel-vite-plugin": "^3.1.0",
     "lit": "^3.3.2",
-    "overtype": "^2.3.10",
+    "overtype": "^2.4.0",
     "tailwindcss": "^4.2.4",
     "vue": "^3.5.33"
   }
