+++ resources/js/modules/markdown-field/behaviors/toolbar/index.ts
@@ -1,4 +1,2 @@
-export {syncToolbarButtonStates} from './active-states';
 export {replaceMarkdownGuideButton} from './guide-link';
 export {toolbarItems} from './items';
-export {syncToolbarKeyboardNavigation} from './keyboard';
